app/requirements/requirements.txt
blinker>=1.8
click>=8.1
cycler>=0.12
flask[async]>=3.0
asgiref>=3.2
itsdangerous>=2.2
Jinja2>=3.1
joblib>=1.4
kiwisolver>=1.4
MarkupSafe>=2.1
matplotlib>=3.9
neurokit2>=0.2
numpy<=1.26.3
onnx>=1.17
onnxruntime>=1.20
packaging>=24
polars>=1.0
pyarrow>=18.0
pillow>=10.3
plotly>=5.20
protobuf>=5.27
pydicom>=2.4
pyparsing>=3.1
python-dateutil>=2.9
pytz>=2024.1
requests>=2.32
scikit-learn>=1.5
scipy>=1.14
six>=1.16
threadpoolctl>=3.5
Werkzeug>=3.1
rlign>=1.1
pandas>=2.2
tqdm>=4.66
flask-session>=0.8
redis[hiredis]>=5.2
celery[redis]>=5.4
gevent>=24.2
gevent-websocket>=0.10
uwsgi>=2.0
webdav4>=0.10
PyYAML>=6.0